+++ src/plonemeeting/portal/core/content/meeting.py
@@ -36,6 +36,10 @@ class IMeeting(model.Schema):
 
     extra_info = RichText(title=_(u"Extra info"), required=False)
 
+    plonemeeting_last_modified = schema.Datetime(
+        title=_(u"Last modification in ia.Delib"), required=True, readonly=True
+    )
+
 
 @implementer(IMeeting)
 class Meeting(Container):