+++ src/plonemeeting/portal/core/tests/test_sync.py
@@ -1,0 +1,7 @@
+import unittest
+
+from plonemeeting.portal.core.testing import PLONEMEETING_PORTAL_CORE_INTEGRATION_TESTING
+
+
+class TestMeetingSynchronization(unittest.TestCase):
+    layer = PLONEMEETING_PORTAL_CORE_INTEGRATION_TESTING