+++ CHANGES.rst
@@ -2,13 +2,32 @@ Changelog
 =========
 
 
-1.0a2 (unreleased)
+1.0a3 (unreleased)
+------------------
+
+- Add basic sync for meeting items & fix localized date conversion
+  [laulaz]
+
+- Added plonemeeting_last_modified to keep track of sync status more easily
+  [odelaere]
+
+- Restrict import action to institutions
+  [laulaz]
+
+- set and format attendees (assembly, assembly excused, assembly absents)
+  [duchenean, boulch]
+
+
+1.0a2 (2019-11-25)
 ------------------
 
 - Fix import meeting form
   [laulaz]
 
 - Fix institution view
+  [laulaz]
+
+- Don't use plone.directives anymore (deprecated)
   [laulaz]